+++ src/main/java/sudoku/solver/model/Grid.java
@@ -3,23 +3,23 @@ package sudoku.solver.model;
 import java.util.Arrays;
 
 /**
- * A mutable Sudoku Grid backed by an array of short[][]
+ * A mutable Sudoku Grid backed by an array of int[][]
  */
 public class Grid implements Cloneable {
-  private short[][] content;
+  private int[][] content;
 
-  public Grid(short[]... lines) {
+  public Grid(int[]... lines) {
     if (lines.length != 9) {
-      lines = new short[9][9];
+      lines = new int[9][9];
     }
     content = lines;
   }
 
-  public short[][] getContent() {
+  public int[][] getContent() {
     return content;
   }
 
-  public short getVal(short x, short y) {
+  public int getVal(int x, int y) {
     return content[y][x];
   }
 
@@ -32,8 +32,8 @@ public class Grid implements Cloneable {
       throw new RuntimeException(cnse);
     }
     // Deep copy of arrays
-    clonedGrid.content = new short[9][];
-    for (short i = 0; i < 9; i++) {
+    clonedGrid.content = new int[9][];
+    for (int i = 0; i < 9; i++) {
       clonedGrid.content[i] = Arrays.copyOf(this.content[i], 9);
     }
     return clonedGrid;
@@ -42,8 +42,8 @@ public class Grid implements Cloneable {
   @Override
   public String toString() {
     StringBuilder builder = new StringBuilder();
-    for (short[] col : content) {
-      for (short cell : col) {
+    for (int[] col : content) {
+      for (int cell : col) {
         builder.append(String.valueOf(cell) + "\t");
       }
       builder.append("\n");
