+++ src/test/java/sudoku/solver/business/RefGrids.java
@@ -7,13 +7,13 @@ import sudoku.solver.model.Grid;
  */
 public class RefGrids {
   public static final Grid AL_ESCARGOT = new Grid(
-      new short[]{1, 0, 0, 0, 0, 7, 0, 9, 0},
-      new short[]{0, 3, 0, 0, 2, 0, 0, 0, 8},
-      new short[]{0, 0, 9, 6, 0, 0, 5, 0, 0},
-      new short[]{0, 0, 5, 3, 0, 0, 9, 0, 0},
-      new short[]{0, 1, 0, 0, 8, 0, 0, 0, 2},
-      new short[]{6, 0, 0, 0, 0, 4, 0, 0, 0},
-      new short[]{3, 0, 0, 0, 0, 0, 0, 1, 0},
-      new short[]{0, 4, 0, 0, 0, 0, 0, 0, 7},
-      new short[]{0, 0, 7, 0, 0, 0, 3, 0, 0});
+      new int[]{1, 0, 0, 0, 0, 7, 0, 9, 0},
+      new int[]{0, 3, 0, 0, 2, 0, 0, 0, 8},
+      new int[]{0, 0, 9, 6, 0, 0, 5, 0, 0},
+      new int[]{0, 0, 5, 3, 0, 0, 9, 0, 0},
+      new int[]{0, 1, 0, 0, 8, 0, 0, 0, 2},
+      new int[]{6, 0, 0, 0, 0, 4, 0, 0, 0},
+      new int[]{3, 0, 0, 0, 0, 0, 0, 1, 0},
+      new int[]{0, 4, 0, 0, 0, 0, 0, 0, 7},
+      new int[]{0, 0, 7, 0, 0, 0, 3, 0, 0});
 }
